requirements.txt
python-telegram-bot==13.15
python-dotenv==1.1.0
google-cloud-dialogflow==2.41.1
vk-api==11.9.9
